Wrote update function and started on start_timer function
