+++ QuickTimer/.idea/misc.xml
@@ -7,6 +7,7 @@
         <entry key="..\:/Users/richyad/Documents/workspace/Kotlin/Utilities/quick-timer/QuickTimer/app/src/main/res/drawable/ic_baseline_play_arrow_24.xml" value="0.15260416666666668" />
         <entry key="..\:/Users/richyad/Documents/workspace/Kotlin/Utilities/quick-timer/QuickTimer/app/src/main/res/drawable/ic_baseline_replay_24.xml" value="0.15260416666666668" />
         <entry key="..\:/Users/richyad/Documents/workspace/Kotlin/Utilities/quick-timer/QuickTimer/app/src/main/res/layout/activity_main.xml" value="0.3333333333333333" />
+        <entry key="..\:/Users/richyad/Documents/workspace/Kotlin/Utilities/quick-timer/QuickTimer/app/src/main/res/layout/second_main.xml" value="0.1" />
       </map>
     </option>
   </component>
