+++ QuickTimer/app/src/main/java/com/codingfanatic/quicktimer/MainActivity.kt
@@ -4,92 +4,82 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
+import android.media.MediaPlayer
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.CountDownTimer
 import android.util.Log
+import android.view.View
 import android.widget.Button
+import android.widget.SeekBar
 import android.widget.TextView
 import com.codingfanatic.quicktimer.databinding.ActivityMainBinding
+import com.codingfanatic.quicktimer.databinding.SecondMainBinding
 import kotlin.math.roundToInt
 import kotlin.math.truncate
 
-class MainActivity : AppCompatActivity() {
+class MainActivity: AppCompatActivity() {
+    var timer_seekBar: SeekBar? = null
+    var timer_textView_two: TextView? =null
+    var timer_button_two: Button? =null
+    var countDownTimer: CountDownTimer? =null
+    var counterIsActive = false
+    var mediaPlayer: MediaPlayer? =null
 
-    companion object {
-        const val TAG = "MainActivity"
-    }
-    private var timeDisplayed: TextView? = null
-    private var timerButton: Button? = null
-    private var someTimer: CountDownTimer? = null
-
-    private var timerStarted = false
-
-    private lateinit var serviceIntent: Intent
-    private lateinit var binding: ActivityMainBinding
-    private var time = 0.0
+    private lateinit var binding: SecondMainBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
+        super.onCreate(savedInstanceState);
+        binding = SecondMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        binding.startStopButton.setOnClickListener{ startStopTimer() }
-        binding.resetButton.setOnClickListener{ resetTimer() }
+        binding.startStopButtonTwo.setOnClickListener{ start_timer() }
+        mediaPlayer = MediaPlayer.create(applicationContext, R.raw.alarm)
 
+        binding.volumeSeekBar.setOnSeekBarChangeListener(object: SeekBar.OnSeekBarChangeListener{
+            override fun onProgressChanged(seekBar: SeekBar?, progress: Int, fromUser: Boolean) {
+                update(progress)
+            }
 
-        serviceIntent = Intent(applicationContext, TimerService::class.java)
-        registerReceiver(updateTime, IntentFilter(TimerService.TIMER_UPDATED))
+            override fun onStartTrackingTouch(seekBar: SeekBar?) {
+                TODO("Not yet implemented")
+            }
+
+            override fun onStopTrackingTouch(seekBar: SeekBar?) {
+                TODO("Not yet implemented")
+            }
+
+        })
+        /*
+            serviceIntent = Intent(applicationContext, TimerService::class.java)
+            registerReceiver(updateTime, IntentFilter(TimerService.TIMER_UPDATED))
+        */
     }
 
-    private fun resetTimer() {
-        stopTimer()
-        time = 0.0
-        binding.timerTextView.text = getTimeStringFromDouble(time)
-    }
+    private fun update(progress: Int){
+        val minutes = progress / 60
+        val seconds = progress % 60
+        val secondsFinal = ""
 
-    private fun startStopTimer() {
-        if(timerStarted){
-            stopTimer()
+        if(seconds <= 9){
+            secondsFinal.plus("0$seconds")
         }
         else{
-            startTimer()
+            secondsFinal.plus(seconds)
+        }
+        timer_seekBar?.setProgress(progress)
+        timer_textView_two?.setText("$minutes:$secondsFinal")
+    }
+
+
+
+
+
+    private fun start_timer(view: View) {
+        if(!counterIsActive){
+            counterIsActive = true;
+            timer_seekBar?.setEnabled(false)
+            //Also set start button text to stop...once we figure out where it is?
         }
     }
-
-    private fun startTimer() {
-        serviceIntent.putExtra(TimerService.TIME_EXTRA, time)
-        startService(serviceIntent)
-        binding.startStopButton.text = "Stop"
-        binding.startStopButton.icon = getDrawable(R.drawable.ic_baseline_pause_24)
-        binding.timerTextView.text = getTimeStringFromDouble(time)
-        timerStarted = true
-    }
-
-    private fun stopTimer() {
-        stopService(serviceIntent)
-        binding.startStopButton.text = "Start"
-        binding.startStopButton.icon = getDrawable(R.drawable.ic_baseline_play_arrow_24)
-        timerStarted = false
-    }
-
-    private val updateTime: BroadcastReceiver = object: BroadcastReceiver() {
-        override fun onReceive(context: Context, intent: Intent) {
-            time = intent.getDoubleExtra(TimerService.TIME_EXTRA, 0.0)
-            binding.timerTextView.text = getTimeStringFromDouble(time)
-        }
-    }
-
-    private fun getTimeStringFromDouble(time: Double): String {
-        val resultInt = truncate(time).toInt()
-        val hours = resultInt % 86400 / 3600
-        val minutes = resultInt % 86400 % 3600 / 60
-        val seconds = resultInt % 86400 % 3600 % 60
-
-        return makeTimeString(hours, minutes, seconds)
-    }
-
-    private fun makeTimeString(hour: Int, min: Int, sec: Int): String = String.format("%02d:%02d:%02d", hour, min, sec)
 }
-
-
