+++ QuickTimer/app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".MainActivityOLD">
 
     <TextView
         android:id="@+id/timer_textView"
