+++ QuickTimer/app/src/main/res/layout/second_main.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+    android:orientation="vertical"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".MainActivity"
+    android:gravity="center">
+
+    <SeekBar
+        android:id="@+id/volumeSeekBar"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:min="0"
+        android:max="100"/>
+
+    <TextView
+        android:id="@+id/timer_texView_two"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="0:30"
+        android:padding="10dp"
+        android:textColor="#000000"
+        android:textSize="50sp"
+        android:textAlignment="center" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/startStop_Button_two"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="START" />
+</LinearLayout>
